Setup droplet

Setup droplet

Transform shape files to mbtiles

Open browser to tippecanoe usage manual

Open browser to tileserver GL documentation

Open browser to GL styling documentation

Reset SSH sessions

Launch tileserver GL

setup_docklet(
  size = "s-2vcpu-4gb-intel",
  region = "tor1",
  image = "docker-20-04",
  tags = c("bccciss", "tileserver"),
  checks = FALSE,
  ...
)

remote_shp_tiles(
  droplet,
  ...,
  source_dir,
  remote_dir = "/tmp/shp",
  skip_upload = FALSE
)

tippecanoe_usage()

tileserver_doc()

styling_doc()

reset_ssh_sessions()

launch_tileserver(droplet, config, styles)

Arguments

size

A character string. Size slug identifier. See analogsea::sizes() for a complete list. Default to s-2vcpu-4gb-intel, large enough to serve layers covering B.C. with fast memory and disk.

region

A character string. The unique slug identifier for the region that you wish to deploy in. See analogsea::regions() for a complete list. Default to tor1 in Canada.

image

A character string. The image ID of a public or private image, or the unique slug identifier for a public image. This image should have Docker already installer. See analogsea::images() for a complete list. Default to docker-20-04.

tags

A vector of tag names to apply to the Droplet after it is created. Tag names can either be existing or new tags. Default to c("bccciss", "tileserver")

checks

A boolean. Check that region is tor1 and image is official Docker. Default to false.

...

additional command line arguments for tippecanoe

droplet

DO droplet

source_dir

Directory containing shape files

remote_dir

Directory where to upload files on remote. Default to /tmp/shp.

skip_upload

A boolean. Skip upload and convert to geojson part. Default FALSE.

config

Local path to a tileserver config file

styles

Local path to Mapbox GL compatible styles (file or dir)

Value

A droplet.

Details

Setup a droplet with necessary tooling to convert shape files to vector tiles and host them using tileserver gl.

Look for shape files in source_dir. Upload source_dir content to droplet remote_dir. Transform shape files to geojson and call tippecanoe with each layers found using ... for parametrization creating tiles.mbtiles. See tippecanoe_usage() for tippecanoe usage documentation.

Launch a tileserver GL docker image. Also stops any already running instances, delete server cache and reload nginx.

If config file is provided, it will be uploaded to server and used to launch tileserver. See https://tileserver.readthedocs.io/en/latest/config.html for details about config